Create roulette with storage object serialized

+++ Roulette.Core/Roulette.Core/Roulette.Core/Contract/IRouletteService.cs
@@ -6,7 +6,7 @@ namespace Roulette.Core.Contract
 {
     public interface IRouletteService
     {
-        int CreateRoulette(string name);
+        Int64 CreateRoulette(string name);
         string OpenRoulette(string id);
     }
 }
